fix: typo in additionalContextProvider file name, move active file inside File context menu
