+++ server/aws-lsp-codewhisperer/src/language-server/agenticChat/context/additionalContextProvider.test.ts
@@ -4,7 +4,7 @@ import { URI } from 'vscode-uri'
 import { TestFeatures } from '@aws/language-server-runtimes/testing'
 import * as assert from 'assert'
 import { AdditionalContextPrompt } from 'local-indexing'
-import { AdditionalContextProvider } from './addtionalContextProvider'
+import { AdditionalContextProvider } from './additionalContextProvider'
 import { getUserPromptsDirectory } from './contextUtils'
 import { LocalProjectContextController } from '../../../shared/localProjectContextController'
 import { workspaceUtils } from '@aws/lsp-core'
