+++ server/aws-lsp-codewhisperer/src/language-server/agenticChat/context/additionalContextProvider.ts
